+++ packages/cardano-services/src/ProgramsCommon/CommonOptionDescriptions.ts
@@ -1,0 +1,5 @@
+export enum CommonOptionDescriptions {
+  LoggerMinSeverity = 'Log level',
+  OgmiosUrl = 'Ogmios URL',
+  RabbitMQUrl = 'RabbitMQ URL'
+}
