+++ packages/cardano-services/src/ProgramsCommon/defaults.ts
@@ -1,0 +1,8 @@
+import { createConnectionObject } from '@cardano-sdk/ogmios';
+
+export const OGMIOS_URL_DEFAULT = (() => {
+  const connection = createConnectionObject();
+  return connection.address.webSocket;
+})();
+
+export const RABBITMQ_URL_DEFAULT = 'amqp://localhost:5672';
