+++ packages/cardano-services/src/ProgramsCommon/options.ts
@@ -1,0 +1,12 @@
+import { LogLevel } from 'bunyan';
+
+/**
+ * Common options for programs:
+ * - HTTP server
+ * - RabbitMQ worker
+ */
+export interface CommonProgramOptions {
+  loggerMinSeverity?: LogLevel;
+  ogmiosUrl?: URL;
+  rabbitmqUrl?: URL;
+}
